Refactor notification helpers
