+++ src/stories/containers/users/delete-account/useDeleteAccount.ts
@@ -6,7 +6,7 @@ import useSWR from 'swr';
 import { GRAPHQL_ENDPOINT } from '../../../../config/endpoints';
 import { useAuthContext } from '../../../../core/context/AuthContext';
 import { fetcher } from '../../../../core/utils/fetcher';
-import { notificationHelper } from '../../../helpers/helpers';
+import { notificationHelper } from '../../../../core/utils/notifications';
 import { LOGIN_REQUEST } from '../../auth/login/login.api';
 import { FETCH_USER_BY_USERNAME } from '../managed-user-profile/managed-user-profile.api';
 import { USERS_DELETE_FROM_ADMIN } from './delete-account.api';
